check db integrity on startup
Looks for version and node documents with an integer version of  -1, indicating
that the save failed before updating can occur.

If the corresponding sample doc doesn't exist or doesn't contain the uuid
version for the version/node doc, then the version is deleted from the
system assuming at least an hour has passed since they were saved.

Otherwise, the version and node docs are updated with the appropriate integer
version.