Fix race condition on setting acls
If the user has changed between  fetching acls from the DB and saving back
updated ACLs, the owner might also be listed in admin/write/read ACLs.
This change ensures that the owner hasn't changed since the ACLs were
retrieved.

Next it to check that the ACLs are consistent  - e.g. no user appears more than
once in the ACLS.