Fix inconsistency in link design document
At the start of the document, it is explicitly said that if a user has
access to data, the user also has access to any linked samples and does
not need read permissions for the samples. This is contradicted later on in
the document. The first design is correct, and is how the link implementation
currently works. Remove the contradictory text.